# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: RISE API
  version: 1.0.0
  description: RISE — Autonomous Incident Remediation System API Specification
servers:
  - url: https://api.rise.internal/api/v1
    description: Internal API Gateway
security:
  - BearerAuth: []
paths:
...
  /knowledge:
    get:
      summary: Browse or search knowledge base entries
      operationId: searchKnowledge
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: service
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Matching knowledge base entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeListEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'

    post:
      summary: Manually create a runbook / knowledge entry
      operationId: createKnowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeCreateRequest'
      responses:
        '201':
          description: Knowledge entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /policies:
    get:
      summary: List risk policies
      operationId: listPolicies
      responses:
        '200':
          description: List of risk policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskPolicyListEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'

    post:
      summary: Define a new risk policy
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreateRequest'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskPolicyEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403RiskPolicyViolation'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /policies/{policy_id}:
    put:
      summary: Update a risk policy (creates a new version)
      operationId: updatePolicy
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyUpdateRequest'
      responses:
        '200':
          description: Policy updated to new version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskPolicyEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
...
  /reports/mttr:
    get:
      summary: Get MTTR report
      operationId: getMttrReport
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: service
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: MTTR analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MttrReportEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /reports/autonomy:
    get:
      summary: Get autonomy resolution percentages
      operationId: getAutonomyReport
      responses:
        '200':
          description: Autonomy percentages report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutonomyReportEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Meta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_POLICY_VIOLATION
            - APPROVAL_EXPIRED
            - ACTION_PLAN_CHANGED
            - RESOURCE_LOCKED
            - INTEGRATION_UNAVAILABLE
            - RATE_LIMITED
            - UNAUTHORIZED
            - NOT_FOUND
        message:
          type: string
        details:
          type: object
    ErrorEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: object
          nullable: true
          default: null
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    KnowledgeEntry:
      type: object
      required: [id, title, content, tags, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        service:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    KnowledgeCreateRequest:
      type: object
      required: [title, content, tags]
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
    KnowledgeListEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeEntry'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    KnowledgeEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/KnowledgeEntry'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    RiskPolicy:
      type: object
      required: [id, action_pattern, environment, risk_tier, requires_approval, max_blast_radius, version]
      properties:
        id:
          type: string
        action_pattern:
          type: string
        environment:
          type: string
        risk_tier:
          type: string
          enum: [low, medium, high, critical]
        requires_approval:
          type: boolean
        max_blast_radius:
          type: integer
        version:
          type: integer
    PolicyCreateRequest:
      type: object
      required: [action_pattern, environment, risk_tier, requires_approval, max_blast_radius]
      properties:
        action_pattern:
          type: string
        environment:
          type: string
        risk_tier:
          type: string
          enum: [low, medium, high, critical]
        requires_approval:
          type: boolean
        max_blast_radius:
          type: integer
    PolicyUpdateRequest:
      type: object
      properties:
        action_pattern:
          type: string
        environment:
          type: string
        risk_tier:
          type: string
          enum: [low, medium, high, critical]
        requires_approval:
          type: boolean
        max_blast_radius:
          type: integer
    RiskPolicyListEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RiskPolicy'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    RiskPolicyEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/RiskPolicy'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
    MttrReport:
      type: object
      required: [avg_mttr_minutes, trend]
      properties:
        avg_mttr_minutes:
          type: number
        trend:
          type: array
          items:
            type: object
    MttrReportEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/MttrReport'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    AutonomyReport:
      type: object
      required: [auto_resolved_pct, human_approved_pct, rejected_pct]
      properties:
        auto_resolved_pct:
          type: number
        human_approved_pct:
          type: number
        rejected_pct:
          type: number
    AutonomyReportEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/AutonomyReport'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
  responses:
    401Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "UNAUTHORIZED"
              message: "Missing or invalid authorization token"
              details: {}
    403RiskPolicyViolation:
      description: Forbidden - risk policy violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "RISK_POLICY_VIOLATION"
              message: "Action blocked by policy engine"
              details: {}
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "NOT_FOUND"
              message: "Requested resource was not found"
              details: {}
...
    422ValidationError:
      description: Unprocessable Entity - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "VALIDATION_ERROR"
              message: "Request body failed schema validation"
              details: {}
...